feat: Importar base de datos en abrir_proyecto_gui
This commit adds the functionality to import a database in the `abrir_proyecto_gui.py` file. It includes a new button `Importar Base de Datos` that prompts the user for a password and then allows them to select an SQL file to import. The database import is handled using the `mysql.exe` executable from XAMPP.